Refactor to remove unnecessary coupling
The bencher should not have access to the thread-pool-crate-abstraction
module to reduce coupling. However, so far it accesses the module to
trigger monomorphization. The ideal way to break the coupling is
probably manual type instantiation. However, rust does not seem to
support that and we have to put up with a type-ascription trick.