Replace rayon by threadpool
The only reference competitor when benchmarking was rayon up to now.
But rayon's thread-pool is not a direct competitor for poolio as it has
a somewhat different API. However, threadpool is a very popular direct
competitor of poolio as the APIs only differs in how threads are
joined which can be easily accommodated by adding a join-method to the
pool-abstraction.