Use distinct pools in http-test
The intention of the test is to show that poolio can make a http-server
multi-threaded. Using poolio also in the client code here is potential
cheating as it is implicitly assumed that poolio works in that regard
(opening the door for Currry's paradox). The other tests actually rule
out the cheating but it is clearer to use a non-poolio pool for the
client.