Encapsulate the thread-joining trick
Due to rust's ownership system it is not possible to join threads
normally if they are embedded in a larger data structure and one has to
work around the issue with a well-known trick. The trick can be be
merged into the usual thread-API resulting in an API that looks pretty
much like the original. This makes the code clearer as it hides the hack.